manage Hotel

+++ src/app/admin/managehotel/page.tsx
@@ -8,6 +8,7 @@ import createHotel from "@/libs/createHotel";
 import updateHotel from "@/libs/updateHotel";
 import deleteHotel from "@/libs/deleteHotel";
 import { HotelItem } from "../../../../interface";
+import Link from "next/link";
 
 export default function ManageHotel() {
   const { data: session } = useSession();
@@ -46,45 +47,49 @@ export default function ManageHotel() {
   }, [hotel]);
 
   return (
-    <main className="w-[100%] flex flex-col items-center space-y-4">
+    <main className="w-[100%] flex flex-row space-y-4">
       {session?.user.role == "admin" ? (
-        <div>
-          <div className="text-xl font-medium">Vaccine Booking</div>
-          <div className="w-fit space-y-2">
-            <div className="text-md text-left text-gray-600">
-              Booking Information
-            </div>
-            <HotelForm
-              hotel={hotel}
-              onHotelChange={(value: HotelItem) => {
-                setHotel(value);
-              }}
-            ></HotelForm>
-          </div>
-          <button
-            name="Book Vaccine"
-            className="block rounded-md bg-sky-600 hover:bg-indigo-600 px-3 py-2 text-white shadow-sm"
-            onClick={makeBooking}
-          >
-            Book Vaccine
-          </button>
-          <button
-            name="Book Vaccine"
-            className="block rounded-md bg-sky-600 hover:bg-indigo-600 px-3 py-2 text-white shadow-sm"
-            onClick={async () => {
-              if (id) {
-                await deleteHotel(session.user.token, id);
+        <div className="w-[70%] m-10 flex flex-col">
+            <div className="text-4xl font-medium py-5">{!id?"Add Hotel":"Edit Hotel"}</div>
+          <div className="">
+            
+              
+              <HotelForm
+                hotel={hotel}
+                onHotelChange={(value: HotelItem) => {
+                  setHotel(value);
+                }}
+              ></HotelForm>
+           <Link href="/admin/allHotels">
+              <button
+                name="Edit Hotel"
+                className="block rounded-full bg-sky-500 px-5 py-2 text-white shadow-sm m-5"
+                onClick={makeBooking}
+              >
+                {id?"Edit Hotel":"Add Hotel"}
+              </button>
+            </Link>
+            {id ?
+            
+              <button
+                name="Delete Hotel"
+                className="block rounded-full bg-red-500 px-5 py-2 text-white shadow-sm m-5"
+                onClick={async () => {
+                  if (id&&confirm("Are you sure you want to delete this booking?")) {
+                    await deleteHotel(session.user.token, id);
 
-                window.location.href = "/hotel";
-              }
-            }}
-          >
-            Delete
-          </button>
+                    window.location.href="/admin/allHotels"
+                  }
+                }}
+              >
+                Delete
+              </button>:""}
+            
+            
+          </div>
+
         </div>
-      ) : (
-        ""
-      )}
+        ) : ("")}
     </main>
   );
 }
